testing_additional_models

+++ clients/main.py
@@ -10,10 +10,18 @@ router = APIRouter(
 
 @router.post("/", response_model=schemas.Client, tags=["Clients"])
 def create_client(client: schemas.Client, db: Session = Depends(get_db)):
-    client_exists = crud.get_client_by_name(db, name=client.name)
+    client_exists = crud.get_client_by_full_name(db, full_name=client.full_name)
     if client_exists:
         raise HTTPException(status_code=400, detail="Client already registered")
     return crud.create_client(db=db, client=client)
+
+
+@router.put("/{client_id}", response_model=schemas.Client,  tags=["Clients"])
+def update_client(client_id: str, client: schemas.Client, db: Session = Depends(get_db)):
+    client_exists = crud.get_client_by_id(db, id=client_id)
+    if not client_exists:
+        raise HTTPException(status_code=400, detail="Client is not registered")
+    return crud.update_client(db=db, id=client_id, client=client)
 
 
 @router.get("/", response_model=list[schemas.Client], tags=["Clients"])
@@ -23,8 +31,8 @@ def read_clients(skip: int = 0, limit: int = 20, db: Session = Depends(get_db)):
 
 
 @router.get("/{client_name}", response_model=schemas.Client, tags=["Clients"])
-def read_client(name: str, db: Session = Depends(get_db)):
-    client = crud.get_client_by_name(db, name=name)
+def read_client(full_name: str, db: Session = Depends(get_db)):
+    client = crud.get_client_by_full_name(db, full_name=full_name)
     if client is None:
         raise HTTPException(status_code=404, detail="Client not found")
     return client
@@ -32,7 +40,9 @@ def read_client(name: str, db: Session = Depends(get_db)):
 
 @router.delete("/{client_id}", response_model=schemas.Client, tags=["Clients"])
 def delete_client(client_id: int, db: Session = Depends(get_db)):
-    client = crud.delete_client(db, client_id=client_id)
-    if client is None:
+    client_exists = crud.get_client_by_id(db, id=client_id)
+    if client_exists:
+        client = crud.delete_client(db, client_id=client_id)
+        return client
+    else:
         raise HTTPException(status_code=404, detail="Client not found")
-    return client
